+++ src/contexts/index.ts
@@ -1,5 +1,8 @@
 export { EdgesContext, useEdgesProperties } from './EdgesContext';
 export { GraphContext, useGraphProperties } from './GraphContext';
+export { GraphMetaContext, useGraphMeta } from './GraphMetaContext';
+export { GraphRegistryContext, useGraphRegistry } from './GraphRegistryContext';
+export { GraphSelectionContext, useGraphSelection } from './GraphSelectionContext';
 export { LayoutContext, useLayoutProperties } from './LayoutContext';
 export { NodesContext, useNodeProperties } from './NodesContext';
 export { PopupsContext, useModals, useToasts } from './PopupsContext';